Add new fale (lesson 11)

+++ lesson10/task6/ILifeCycle.java
@@ -1,7 +1,8 @@
 package homework.lesson10.task6;
 
 public interface ILifeCycle {
-    void init(double x, double y, GameContext ctx);
+    void init(int x, int y, GameContext ctx);
     void update(GameContext ctx);
     void destroy();
+
 }
